Relocated GlobalState to server package.
