+++ server/file_finder.go
@@ -7,8 +7,6 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-
-	"github.com/sfbrigade/sfsbook/setup"
 )
 
 //go:generate go run ../generator/tool/generateresources.go -output embedded_resources.go -prefix ../site/ ../site
@@ -23,24 +21,24 @@ import (
 type Serve interface {
 	// The desired content is available as a file.
 	ServeForStream(reader io.Reader, w http.ResponseWriter, req interface{})
-	ServeForString(s string, w http.ResponseWriter, req interface{}) 
+	ServeForString(s string, w http.ResponseWriter, req interface{})
 }
 
 // TODO(rjk): This is wrong. It should really be that the Global state is constructed by
 // including a FileFinder instance. And then the global state would provide the
 // methods of the FileFinder. And all of this should get relocated to server.
-type FileFinder setup.GlobalState
+type FileFinder GlobalState
 
-func MakeFileFinder(global *setup.GlobalState) *FileFinder {
+func MakeFileFinder(global *GlobalState) *FileFinder {
 	// Types have to align. This isn't very good code.
 	return (*FileFinder)(global)
 }
 
-// Given an Url, 
+// Given an Url,
 func (ff *FileFinder) StreamOrString(upath string, serve Serve, w http.ResponseWriter, req interface{}) error {
 	fpath := filepath.Join(ff.Sitedir, upath)
 	log.Println(upath, fpath)
-		
+
 	if _, err := os.Stat(fpath); err != nil {
 		res, ok := Resources[upath]
 		if !ok {
@@ -55,7 +53,7 @@ func (ff *FileFinder) StreamOrString(upath string, serve Serve, w http.ResponseW
 	f, err := os.Open(fpath)
 	if err != nil {
 		log.Println("problem opening file", fpath, err)
-		return fmt.Errorf("file %s missing from site: %v", upath,  err)
+		return fmt.Errorf("file %s missing from site: %v", upath, err)
 	}
 	defer f.Close()
 	serve.ServeForStream(f, w, req)
